+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to
 
 ### Fixed
 
+- 🐛(frontend) preserve page titles when adding an emoji
 - 🐛(frontend) refresh pins after document deletion and restoration
 - 🐛(frontend) redirect homepage to login when homepage feat is disabled #2521
 - 🐛(backend) ignore CSPs for API docs in development